// src/app/03_SIR.Produccion.Reportes/pages/cuota/interfaces/DWCorte.interface.ts
export interface DWCorte {
    fecha:          string;
    lote:           number;
    customer?:      string;
    condition:      string;
    marca:          number;
    material:       number;
    codigo:         string;
    producto:       string;
    piezas:         number;
    peso:           number;
}